Uniformly sort diagnostics to the end of draws output
